BUG 347755: Use qreal for constants defined as 0.30 or 0.05

+++ libs/image/kis_liquify_transform_worker.cpp
@@ -431,7 +431,7 @@ void KisLiquifyTransformWorker::run(KisPaintDeviceSP device)
 
 QRect KisLiquifyTransformWorker::approxChangeRect(const QRect &rc)
 {
-    const int margin = 0.05;
+    const qreal margin = 0.05;
 
     /**
      * Here we just return the full area occupied by the transformed grid.
